Rough outline for a GTK+ window and strike-temp.c
Most of this getting the Makefile to compile the code and
building a primative GTK+ window that actually does something.
All it does now is add and subtract -- but that will change soon.

strike-temp.c is the start of a routine to calculate the required
temperature for strike water. Currently it can do a single infusion.